# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KAVACH AI",
    "description": "India's AI-Powered Digital Safety Shield \u2014 Defeating Counterfeiting, Fraud & Digital Arrest Scams. Multi-agent AI platform for citizens, law enforcement, and financial institutions.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/analytics/overview": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Overview",
        "description": "Get platform-wide overview statistics for the dashboard.",
        "operationId": "get_overview_api_v1_analytics_overview_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/threats": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Threat Distribution",
        "description": "Get the distribution of threat types for charts.",
        "operationId": "get_threat_distribution_api_v1_analytics_threats_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreatDistribution"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/timeline": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Timeline",
        "description": "Get temporal trend data for reports.",
        "operationId": "get_timeline_api_v1_analytics_timeline_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "daily",
              "title": "Period"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/geospatial": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Geospatial Data",
        "description": "Get geographic crime intelligence data for maps.",
        "operationId": "get_geospatial_data_api_v1_analytics_geospatial_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeospatialData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/fraud-network": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get Fraud Network",
        "description": "Get fraud network graph data for visualization.",
        "operationId": "get_fraud_network_api_v1_analytics_fraud_network_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FraudNetworkData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FraudEdge": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          },
          "transaction_count": {
            "type": "integer",
            "title": "Transaction Count"
          },
          "total_amount": {
            "type": "number",
            "title": "Total Amount"
          },
          "edge_type": {
            "type": "string",
            "title": "Edge Type"
          }
        },
        "type": "object",
        "required": [
          "source",
          "target",
          "weight",
          "transaction_count",
          "total_amount",
          "edge_type"
        ],
        "title": "FraudEdge",
        "description": "Edge in the fraud network graph."
      },
      "FraudNetworkData": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/FraudNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/FraudEdge"
            },
            "type": "array",
            "title": "Edges"
          },
          "communities": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Communities"
          },
          "central_nodes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Central Nodes"
          },
          "risk_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Risk Summary"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "edges",
          "communities",
          "central_nodes",
          "risk_summary"
        ],
        "title": "FraudNetworkData",
        "description": "Fraud network graph data for visualization."
      },
      "FraudNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "node_type": {
            "type": "string",
            "title": "Node Type"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "is_flagged": {
            "type": "boolean",
            "title": "Is Flagged"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "node_type",
          "risk_score",
          "is_flagged"
        ],
        "title": "FraudNode",
        "description": "Node in the fraud network graph."
      },
      "GeospatialData": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/GeospatialPoint"
            },
            "type": "array",
            "title": "Points"
          },
          "hotspots": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Hotspots"
          },
          "state_summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "State Summary"
          },
          "total_incidents": {
            "type": "integer",
            "title": "Total Incidents"
          }
        },
        "type": "object",
        "required": [
          "points",
          "hotspots",
          "state_summary",
          "total_incidents"
        ],
        "title": "GeospatialData",
        "description": "Geographic crime intelligence data."
      },
      "GeospatialPoint": {
        "properties": {
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "threat_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threat Type"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "count",
          "severity"
        ],
        "title": "GeospatialPoint",
        "description": "Single geographic data point."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OverviewStats": {
        "properties": {
          "total_users": {
            "type": "integer",
            "title": "Total Users"
          },
          "total_reports": {
            "type": "integer",
            "title": "Total Reports"
          },
          "total_scans": {
            "type": "integer",
            "title": "Total Scans"
          },
          "total_cases": {
            "type": "integer",
            "title": "Total Cases"
          },
          "active_alerts": {
            "type": "integer",
            "title": "Active Alerts"
          },
          "threats_blocked": {
            "type": "integer",
            "title": "Threats Blocked"
          },
          "total_amount_saved": {
            "type": "number",
            "title": "Total Amount Saved"
          },
          "scams_detected_today": {
            "type": "integer",
            "title": "Scams Detected Today"
          },
          "reports_by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Reports By Type"
          },
          "reports_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Reports By Status"
          },
          "reports_by_severity": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Reports By Severity"
          },
          "scans_by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Scans By Type"
          }
        },
        "type": "object",
        "required": [
          "total_users",
          "total_reports",
          "total_scans",
          "total_cases",
          "active_alerts",
          "threats_blocked",
          "total_amount_saved",
          "scams_detected_today",
          "reports_by_type",
          "reports_by_status",
          "reports_by_severity",
          "scans_by_type"
        ],
        "title": "OverviewStats",
        "description": "Platform-wide overview statistics."
      },
...
      "ThreatDistribution": {
        "properties": {
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Labels"
          },
          "values": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Values"
          },
          "percentages": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Percentages"
          }
        },
        "type": "object",
        "required": [
          "labels",
          "values",
          "percentages"
        ],
        "title": "ThreatDistribution",
        "description": "Threat type distribution for charts."
      },
      "TimelineData": {
        "properties": {
          "period": {
            "type": "string",
            "title": "Period"
          },
          "data_points": {
            "items": {
              "$ref": "#/components/schemas/TimelineDataPoint"
            },
            "type": "array",
            "title": "Data Points"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "period",
          "data_points",
          "total"
        ],
        "title": "TimelineData",
        "description": "Temporal trend data."
      },
      "TimelineDataPoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "date",
          "count"
        ],
        "title": "TimelineDataPoint",
        "description": "Single data point in a timeline."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}